feat(Note): Add the note class that contains methods to interact with the note content
This is ready to be integrated as part of the refactor work + TDD approach conducted.
I will be now moving to the NoteManager class: in order to replace
wikilinks, note must be aware of their siblings which I don't believe
should belong in the noteclass, but a different class such a LinkManager
or NoteManager.