refactor([WIP]): Removing the event emmiter and replacing the linkManager by the NotesManager
It feels like we're doing the opposite of what we should do: Note have to notifiy the LinkManager that they exist, but they also have a closure on the link manager.
Really, it should be the other way around. The LinkManager should be a NotesManager, in charge of creating the notes. The NotesManager could then go into each notes and get the link out, replacing them with the right link.

It just felt like the code inheritance was the other way around.
This would reduce drastically the public API at the same time